chore: move tests to `finalization_test.go` remove unneeded `time.Sleep`
