+++ lib/grandpa/finalization_integration_test.go
@@ -14,134 +14,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func Test_FinalizationHandler_runEphemeralServices(t *testing.T) {
-	t.Parallel()
-
-	mockingErr := errors.New("mocked error")
-
-	tests := map[string]struct {
-		callHandlerStop           bool
-		createFinalizationHandler func(*gomock.Controller) *finalizationHandler
-		wantErr                   error
-	}{
-		"voting_round_finalisation_engine_finishes_successfully": {
-			createFinalizationHandler: func(ctrl *gomock.Controller) *finalizationHandler {
-				builder := func() (engine ephemeralService, voting ephemeralService) {
-					mockVoting := NewMockephemeralService(ctrl)
-					mockVoting.EXPECT().Run().DoAndReturn(func() error {
-						time.Sleep(3 * time.Second)
-						return nil
-					})
-
-					mockEngine := NewMockephemeralService(ctrl)
-					mockEngine.EXPECT().Run().DoAndReturn(func() error {
-						time.Sleep(4 * time.Second)
-						return nil
-					})
-
-					return mockEngine, mockVoting
-				}
-
-				return &finalizationHandler{
-					newServices: builder,
-					stopCh:      make(chan struct{}),
-					handlerDone: make(chan struct{}),
-				}
-			},
-		},
-
-		"voting_round_fails_should_stop_engine_service": {
-			wantErr: mockingErr,
-			createFinalizationHandler: func(ctrl *gomock.Controller) *finalizationHandler {
-				builder := func() (engine ephemeralService, voting ephemeralService) {
-					const failTime = 2 * time.Second
-
-					mockVoting := NewMockephemeralService(ctrl)
-					mockVoting.EXPECT().Run().DoAndReturn(func() error {
-						time.Sleep(failTime)
-						return mockingErr
-					})
-					mockVoting.EXPECT().Stop().Return(nil)
-
-					// once the voting round fails the finalisation handler
-					// should be awere of the error and call the stop method from
-					// the engine which will release the start method from engine service
-					engineStopCh := make(chan struct{})
-					mockEngine := NewMockephemeralService(ctrl)
-					mockEngine.EXPECT().Run().DoAndReturn(func() error {
-						<-engineStopCh
-						return nil
-					})
-					mockEngine.EXPECT().Stop().DoAndReturn(func() error {
-						close(engineStopCh)
-						return nil
-					})
-
-					return mockEngine, mockVoting
-				}
-
-				return &finalizationHandler{
-					newServices: builder,
-					stopCh:      make(chan struct{}),
-					handlerDone: make(chan struct{}),
-				}
-			},
-		},
-
-		"engine_fails_should_stop_voting_round_service": {
-			wantErr: mockingErr,
-			createFinalizationHandler: func(ctrl *gomock.Controller) *finalizationHandler {
-				builder := func() (engine ephemeralService, voting ephemeralService) {
-					const failTime = 2 * time.Second
-
-					mockEngine := NewMockephemeralService(ctrl)
-					mockEngine.EXPECT().Run().DoAndReturn(func() error {
-						time.Sleep(failTime)
-						return mockingErr
-					})
-					mockEngine.EXPECT().Stop().Return(nil)
-
-					// once the finalisation engine fails the finalisation handler
-					// should be awere of the error and call the stop method from the
-					// voting round which will release the start method from voting round service
-					votingStopChannel := make(chan struct{})
-					mockVoting := NewMockephemeralService(ctrl)
-					mockVoting.EXPECT().Run().DoAndReturn(func() error {
-						<-votingStopChannel
-						return nil
-					})
-					mockVoting.EXPECT().Stop().DoAndReturn(func() error {
-						close(votingStopChannel)
-						return nil
-					})
-
-					return mockEngine, mockVoting
-				}
-
-				return &finalizationHandler{
-					newServices: builder,
-					stopCh:      make(chan struct{}),
-					handlerDone: make(chan struct{}),
-				}
-			},
-		},
-	}
-
-	for tname, tt := range tests {
-		tt := tt
-
-		t.Run(tname, func(t *testing.T) {
-			t.Parallel()
-
-			ctrl := gomock.NewController(t)
-			finalizationHandler := tt.createFinalizationHandler(ctrl)
-
-			err := finalizationHandler.runEphemeralServices()
-			require.ErrorIs(t, err, tt.wantErr)
-		})
-	}
-}
-
 func Test_FinalizationHandler_Stop_ShouldHalt_Services(t *testing.T) {
 	t.Parallel()
